user_service/requirements.txt
aiohttp>=3.6
sqlalchemy>1.3
Alembic>1.4
python-dotenv>=0.14
passlib>1.7
argon2-cffi>=20.0
aiopg>=1.0
aioredis>=1.3
aiohttp_session>=2.9
pyjwt>=1.7
python-dotenv>=0.14
aiohttp-swagger>=1.0